+++ RELEASE_NOTES.md
@@ -1,9 +1,11 @@
 # Release Notes
 
-## 5.18.4-alpha - tbd
+## 5.19-alpha - 2019-12-15
 
+* EHNANCEMENT: `Fake.DotNet.Cli` Add timeout field to kill the process after the given timeout - https://github.com/fsharp/FAKE/pull/2425
+* EHNANCEMENT: `Fake.Core.Target` Add notification when a fake worker is idle - https://github.com/fsharp/FAKE/pull/2425
 * EHNANCEMENT: Use SPDX license expressions in the nuspec files, thanks @teo-tsirpanis - https://github.com/fsharp/FAKE/pull/2403
-* ENHANCEMENT: Update bundled paket - https://github.com/fsharp/FAKE/pull/2425
+* ENHANCEMENT: `fake-cli` Update bundled paket - https://github.com/fsharp/FAKE/pull/2425
 * ENHANCEMENT: `Fake.DotNet.Cli` Suppress copyright messages when invoking `dotnet build`, thanks @objectx - https://github.com/fsharp/FAKE/pull/2436
 * EHNANCEMENT: `Fake.Installer.Squirrel` Add additional Squirrel parameters, thanks @pchinery - https://github.com/fsharp/FAKE/pull/2431
 * BUGFIX: `Fake.Tools.Rsync` Fix typo in Rsync.fs options, thanks @rmunn - https://github.com/fsharp/FAKE/pull/2432